cherry pick 553e8f9 from https://gitee.com/atrotro/third_party_ffmpeg/pulls/297 【修改说明】：修复ffmpeg问题

+++ libavformat/aacdec.c
@@ -220,7 +220,7 @@ static void adts_aac_get_duration(AVFormatContext *s, AVStream *st)
     }
 
     st->codecpar->channels = channel;
-    st->codecpar->sample_rate = sr;
+    st->codecpar->sample_rate = (int)sr;
     avpriv_set_pts_info(st, 64, 1, st->codecpar->sample_rate);
 
     int frame_size = 0;
